Refactor details out of PrevalenceControls
