+++ src/components/calculator/PrevalenceControls.tsx
@@ -1,19 +1,14 @@
 import i18n from 'i18n'
 import countries from 'i18n-iso-countries'
-import { isNumber } from 'lodash'
 import React, { useEffect, useState } from 'react'
-import { Card, Form, InputGroup, ToggleButton } from 'react-bootstrap'
+import { ToggleButton } from 'react-bootstrap'
 import { Typeahead } from 'react-bootstrap-typeahead'
 import { Trans, useTranslation } from 'react-i18next'
-import { BsDash } from 'react-icons/bs'
 
 import ControlLabel from './controls/ControlLabel'
-import { ControlledExpandable } from 'components/Expandable'
-import {
-  CalculatorData,
-  calculateLocationPersonAverage,
-  calculateLocationReportedPrevalence,
-} from 'data/calculate'
+import { LocationPrevalenceDetails } from './prevalence/LocationPrevalenceDetails'
+import { ManualPrevalenceDetails } from './prevalence/ManualPrevalenceDetails'
+import { CalculatorData } from 'data/calculate'
 import { Locations, PrevalenceDataDate } from 'data/location'
 import 'components/calculator/styles/PrevalenceControls.scss'
 
@@ -28,112 +23,6 @@ const isFilled = (val: string): boolean => {
 
 const isTopLocation = (val: string): boolean => {
   return isFilled(val) && !!Locations[val]
-}
-
-const PrevalenceFieldStatic: React.FunctionComponent<{
-  id: string
-  label: string
-  value: string | number
-  unit?: string
-}> = ({ id, label, value, unit }): React.ReactElement => {
-  return (
-    <div id={id}>
-      {label}: {typeof value === 'number' ? value.toLocaleString() : value}
-      {unit}
-    </div>
-  )
-}
-
-const PrevalenceField: React.FunctionComponent<{
-  id: string
-  label: string
-  value: string | number
-  unit?: string
-  setter: (newValue: string) => void
-  inputType: string
-  max?: number
-  min?: number
-  helpText?: string
-  className?: string
-}> = ({
-  id,
-  label,
-  value,
-  setter,
-  unit,
-  inputType,
-  max,
-  min,
-  helpText,
-  className,
-}): React.ReactElement => {
-  let body: React.ReactElement = (
-    <Form.Control
-      className={'form-control form-control-lg col-md-3 col-lg-6 ' + className}
-      type={inputType}
-      value={value}
-      readOnly={false}
-      onChange={(e) => {
-        if (isNumber(max) || isNumber(min)) {
-          let newValue = Number.parseFloat(e.target.value)
-          newValue = isNumber(max) && newValue > max ? max : newValue
-          newValue = isNumber(min) && newValue < min ? min : newValue
-          setter(newValue.toString())
-        } else {
-          setter(e.target.value)
-        }
-      }}
-    />
-  )
-  if (unit) {
-    body = (
-      <InputGroup className="mb-3">
-        {body}
-        <InputGroup.Append>
-          <InputGroup.Text>%</InputGroup.Text>
-        </InputGroup.Append>
-      </InputGroup>
-    )
-  }
-  return (
-    <Form.Group controlId={id} className="mb-3">
-      <Form.Label>{label}</Form.Label>
-      {body}
-      {helpText && (
-        <Form.Text id={id + 'HelpText'} muted>
-          {helpText}
-        </Form.Text>
-      )}
-    </Form.Group>
-  )
-}
-
-export const PrevalenceResult = (props: {
-  data: CalculatorData
-}): React.ReactElement => {
-  return (
-    <Card className="prevalence-result">
-      <Card.Body>
-        <div>
-          <Trans>calculator.prevalence.reported_prevalence</Trans>:{' '}
-          {(
-            (calculateLocationReportedPrevalence(props.data) || 0) * 100
-          ).toFixed(2)}
-          %
-        </div>
-        <div>
-          <strong>
-            <Trans>calculator.prevalence.adjusted_prevalence</Trans>:{' '}
-            {(
-              ((calculateLocationPersonAverage(props.data) || 0) * 100) /
-              1e6
-            ).toFixed(2)}
-            %
-          </strong>
-        </div>
-      </Card.Body>
-    </Card>
-  )
 }
 
 export const PrevalenceControls: React.FunctionComponent<{
@@ -341,149 +230,21 @@ export const PrevalenceControls: React.FunctionComponent<{
         </ToggleButton>
       </span>
       {isManualEntryCurrently ? (
-        <div id="prevalence-details">
-          <span id="details-header">
-            <BsDash /> {t('calculator.prevalence.details_header')}
-          </span>
-          <PrevalenceResult data={data} />
-          <PrevalenceField
-            id="reported-cases"
-            label={t('calculator.prevalence.last_week_cases')}
-            value={data.casesPastWeek || 0}
-            setter={(value) =>
-              setter({ ...data, casesPastWeek: parseInt(value || '') })
-            }
-            inputType="number"
-            className="hide-number-buttons"
-          />
-          <PrevalenceField
-            id="population"
-            label={t('calculator.prevalence.population')}
-            value={data.population}
-            setter={(value) => setter({ ...data, population: value })}
-            inputType="text"
-            className="hide-number-buttons"
-          />
-          <PrevalenceField
-            id="percent-increase"
-            label={t('calculator.prevalence.percent_increase_in_cases')}
-            value={data.casesIncreasingPercentage}
-            unit="%"
-            setter={(value) => {
-              setter({ ...data, casesIncreasingPercentage: Number(value) })
-            }}
-            inputType="number"
-            min={0}
-            className="hide-number-buttons"
-          />
-          <PrevalenceField
-            id="positive-test-rate"
-            label={t('calculator.prevalence.positive_case_percentage')}
-            value={
-              data.positiveCasePercentage
-                ? data.positiveCasePercentage.toString()
-                : ''
-            }
-            unit="%"
-            setter={(value) => {
-              setter({ ...data, positiveCasePercentage: Number(value) })
-            }}
-            inputType="number"
-            max={100}
-            min={0}
-            className="hide-number-buttons"
-          />
-        </div>
-      ) : (
-        <ControlledExpandable
+        <ManualPrevalenceDetails
           id="prevalence-details"
+          data={data}
+          setter={setter}
+        />
+      ) : (
+        <LocationPrevalenceDetails
+          id="prevalence-details"
+          data={data}
           header={t('calculator.prevalence.details_header')}
-          headerClassName={isManualEntryCurrently ? 'd-none' : ''}
           open={detailsOpen}
           setter={setDetailsOpen}
-        >
-          <PrevalenceResult data={data} />
-
-          <PrevalenceFieldStatic
-            id="reported-cases"
-            label={t('calculator.prevalence.last_week_cases')}
-            value={data.casesPastWeek || 0}
-          />
-          <PrevalenceFieldStatic
-            id="population"
-            label={t('calculator.prevalence.population')}
-            value={data.population}
-          />
-          {locationSet && data.casesIncreasingPercentage === 0 ? (
-            <div>{t('calculator.prevalence.cases_stable_or_decreasing')}</div>
-          ) : (
-            <PrevalenceFieldStatic
-              id="percent-increase"
-              label={t('calculator.prevalence.percent_increase_in_cases')}
-              value={data.casesIncreasingPercentage}
-              unit="%"
-            />
-          )}
-          <PrevalenceFieldStatic
-            id="positive-test-rate"
-            label={t('calculator.prevalence.positive_case_percentage')}
-            value={
-              data.positiveCasePercentage === null
-                ? 'no data available'
-                : data.positiveCasePercentage.toString()
-            }
-            unit="%"
-          />
-          {!locationSet ? null : (
-            <>
-              <div>
-                <em>
-                  <Trans>calculator.prevalence.data_last_updated</Trans>:{' '}
-                  {PrevalenceDataDate}
-                </em>
-              </div>
-              <div>
-                <p className="mt-3">
-                  <Trans i18nKey="calculator.prevalence_info_source_information">
-                    Prevalence data consolidated from {}
-                    <a
-                      href="https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data"
-                      target="_blank"
-                      rel="noreferrer"
-                    >
-                      Johns Hopkins CSSE
-                    </a>{' '}
-                    (reported cases), {}
-                    <a
-                      href="https://apidocs.covidactnow.org/"
-                      target="_blank"
-                      rel="noreferrer"
-                    >
-                      Covid Act Now
-                    </a>{' '}
-                    (US positive test rates), {}
-                    <a
-                      href="https://ourworldindata.org/coronavirus-testing#testing-for-covid-19-background-the-our-world-in-data-covid-19-testing-dataset"
-                      target="_blank"
-                      rel="noreferrer"
-                    >
-                      Our World in Data
-                    </a>{' '}
-                    (international positive test rates), and {}
-                    <a
-                      href="https://covid19.geo-spatial.org/despre"
-                      target="_blank"
-                      rel="noreferrer"
-                    >
-                      Coronavirus COVID-19 România
-                    </a>{' '}
-                    (Romania reported cases).
-                  </Trans>
-                </p>
-              </div>
-            </>
-          )}
-        </ControlledExpandable>
+          hide={isManualEntryCurrently}
+          locationSet={locationSet}
+        />
       )}
     </React.Fragment>
   )
