+++ src/components/calculator/prevalence/PrevalenceResult.tsx
@@ -1,0 +1,39 @@
+import React from 'react'
+import { Card } from 'react-bootstrap'
+import { Trans } from 'react-i18next'
+
+import {
+  CalculatorData,
+  calculateLocationPersonAverage,
+  calculateLocationReportedPrevalence,
+} from 'data/calculate'
+
+export const PrevalenceResult = (props: {
+  data: CalculatorData
+}): React.ReactElement => {
+  return (
+    <Card className="prevalence-result">
+      <Card.Body>
+        <div>
+          <Trans>calculator.prevalence.reported_prevalence</Trans>:{' '}
+          {(
+            (calculateLocationReportedPrevalence(props.data) || 0) * 100
+          ).toFixed(2)}
+          %
+        </div>
+        <div>
+          <strong>
+            <Trans>calculator.prevalence.adjusted_prevalence</Trans>:{' '}
+            {(
+              ((calculateLocationPersonAverage(props.data) || 0) * 100) /
+              1e6
+            ).toFixed(2)}
+            %
+          </strong>
+        </div>
+      </Card.Body>
+    </Card>
+  )
+}
+
+export default PrevalenceResult
