+++ src/components/calculator/styles/PrevalenceControls.scss
@@ -23,7 +23,7 @@
   color: $gray-500 !important;
 }
 
-#details-header {
+.details-header {
   font-weight: bold;
   display: block;
   padding-bottom: 0.5em;
